tests: Move non-editor test content to its own file
To make it easier to run some of the unit tests on the very
spectrum itself!
